Expose a QuorumMasterAddr method
At the moment, if any sentinel is reachable and able to provide a master
address, the sentinelFailover accepts it. We'd like to know that the
sentinel providing the master address has a quorum. This should reduce
the likelihood of a sentinel in a minority partition with a previous
master providing an invalid master address.